SubjectIcon with mono option * removed SubjectMonoIcon
