--- src/components/subject-mono-icons/SubjectMonoIcon.jsx
@@ -1,102 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-
-const TYPE = {
-  ACCOUNTANCY: 'accountancy',
-  ADMINISTRATION: 'administration',
-  AGRICULTURAL: 'agricultural',
-  ALGEBRA: 'algebra',
-  ALL: 'all',
-  ARABIC: 'arabic',
-  ART: 'art',
-  ARTMUSIC: 'artmusic',
-  BELARUS: 'belarus',
-  BIOLOGY: 'biology',
-  BUSINESS: 'business',
-  CATALA: 'catala',
-  CHEMISTRY: 'chemistry',
-  CHINESE: 'chinese',
-  ECONOMICS: 'economics',
-  EGZAM: 'egzam',
-  ENGLISH: 'english',
-  ENTREPRENEURSHIP: 'entrepreneurship',
-  ENVIRONMENT: 'environment',
-  ETHICS: 'ethics',
-  EUSKARA: 'euskara',
-  FIRST_AID: 'first-aid',
-  FRENCH: 'french',
-  GALEGO: 'galego',
-  GEOGRAPHY: 'geography',
-  GEOLOGY: 'geology',
-  GEOMETRY: 'geometry',
-  GERMAN: 'german',
-  GRAMMAR: 'grammar',
-  HEALTH: 'health',
-  HISTORY: 'history',
-  INDIA_LANG: 'india-lang',
-  INDONESIAN_LANG: 'indonesian-lang',
-  INFORMATICS: 'informatics',
-  ITALIAN: 'italian',
-  JAPANESE: 'japanese',
-  KAZACH: 'kazach',
-  KOREAN: 'korean',
-  LANGUAGE: 'language',
-  LATIN: 'latin',
-  LAW: 'law',
-  LIFE_SCIENCE: 'life-science',
-  LITERATURE: 'literature',
-  LOGIC: 'logic',
-  MATHEMATICS: 'mathematics',
-  MUSIC: 'music',
-  NIGERIAN_LANG: 'nigerian-lang',
-  OTHERLANGUAGES: 'otherlanguages',
-  OTHERS: 'others',
-  PEDAGOGICS: 'pedagogics',
-  PHILOSOPHY: 'philosophy',
-  PHYSICAL_EDUCATION: 'physical-education',
-  PHYSICS: 'physics',
-  POLITICS: 'politics',
-  PSYCHOLOGY: 'psychology',
-  RELIGION: 'religion',
-  RPA_LANG: 'rpa-lang',
-  RUSSIAN: 'russian',
-  SCIENCE: 'science',
-  SECURITY: 'security',
-  SKILLS: 'skills',
-  SOCIAL_SCIENCE: 'social-science',
-  SOCIOLOGY: 'sociology',
-  SPANISH: 'spanish',
-  STATISTICS: 'statistics',
-  TECHNOLOGY: 'technology',
-  TOURISM: 'tourism',
-  TRAFFIC: 'traffic',
-  UKRAINIAN: 'ukrainian',
-  UKRAINIAN_LITERATURE: 'ukrainian-literature',
-  WOS: 'wos'
-};
-
-const SIZE = {
-  SMALL: 'small',
-  MEDIUM: 'medium',
-  NORMAL: 'normal'
-};
-
-const SubjectMonoIcon = ({type, size = SIZE.NORMAL}) => {
-  const iconClass = classNames('sg-subject-icon', {
-    [`sg-subject-icon--${size}`]: size !== SIZE.NORMAL
-  });
-  const iconType = `#icon-subject-mono-${type}`;
-
-  return <svg className={iconClass}>
-    <use xlinkHref={iconType}></use>
-  </svg>;
-};
-
-SubjectMonoIcon.propTypes = {
-  type: PropTypes.oneOf(Object.values(TYPE)).isRequired,
-  size: PropTypes.oneOf(Object.values(SIZE))
-};
-
-export default SubjectMonoIcon;
-export {TYPE, SIZE};
