--- src/components/subject-mono-icons/pages/subject-mono-icons-interactive.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import SubjectMonoIcon, {TYPE, SIZE} from '../SubjectMonoIcon';
-import DocsActiveBlock from 'components/DocsActiveBlock';
-
-const SubjectMonoIcons = () => {
-  const settings = [
-    {
-      name: 'type',
-      values: TYPE,
-      required: true
-    },
-    {
-      name: 'size',
-      values: SIZE
-    }
-  ];
-
-  return <div>
-    <DocsActiveBlock settings={settings}>
-      <SubjectMonoIcon type={TYPE.ENTREPRENEURSHIP}/>
-    </DocsActiveBlock>
-  </div>;
-};
-
-export default SubjectMonoIcons;
